prometheus-exporter/requirements.txt
# Prometheus Exporter Dependencies
pyyaml>=6.0
